CASTOR-4976 tapeserverd sometimes wrongly counts transfered bytes
Wrong type led to truncation (and corruption) for file sizes bigger that 2GB. (int, where uint64_t should have been used).

+++ castor/tape/tapeserver/daemon/TapeWriteTask.hpp
@@ -65,7 +65,7 @@ public:
   /**
    * @return the size of the file in byte
    */
-  virtual int fileSize();
+  virtual uint64_t fileSize();
     
   /**
    * Main execution routine
